replace the old rllib agent
